Type change

+++ src/mainCode/Ball.java
@@ -7,13 +7,13 @@ import java.awt.Point;
 import java.awt.geom.Point2D;
 
 public class Ball {
-	private int x;
-	private int y;
-	private int diameter;
+	private double x;
+	private double y;
+	private double diameter;
 	private Color color;
 	private double slope;
 	
-	public Ball(int x, int y, int diameter, Color color){
+	public Ball(double x, double y, double diameter, Color color){
 		this.x = x;
 		this.y = y;
 		this.diameter = diameter;
@@ -38,7 +38,7 @@ public class Ball {
 	public void draw(Graphics g) {
 		//Graphics2D g2 = (Graphics2D) g;
 		g.setColor(color);
-		g.fillOval(x, y, diameter, diameter);
+		g.fillOval((int) x, (int) y, (int) diameter, (int) diameter);
 	}
 	
 	public void getIcon () {
@@ -57,7 +57,7 @@ public class Ball {
 		return false;
 	}
 	
-	public void setPosition(int x, int y) {
+	public void setPosition(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
@@ -76,23 +76,23 @@ public class Ball {
 		return 0;
 	}
 
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 
-	public int getDiameter() {
+	public double getDiameter() {
 		return diameter;
 	}
 }
